exploration to simulate bad scrolling memory usage
See https://github.com/yi-editor/yi/issues/105

ideally purely, but if not at least be able to have a compilable program
that reproduces the bug reliably with a known ghc source file.

These are notes in that direction.

The end goal is to make tests that will fail if scrolling down for
instance becomes too slow